refactoring

+++ schema-loader/src/main/java/com/scalar/db/schemaloader/command/ConfigFileBasedCommand.java
@@ -1,11 +1,11 @@
 package com.scalar.db.schemaloader.command;
 
 import com.scalar.db.config.DatabaseConfig;
-import com.scalar.db.schemaloader.command.CassandraCommand.CompactStrategy;
-import com.scalar.db.schemaloader.command.CassandraCommand.ReplicationStrategy;
 import com.scalar.db.schemaloader.core.SchemaOperator;
 import com.scalar.db.schemaloader.schema.SchemaParser;
 import com.scalar.db.storage.cassandra.CassandraAdmin;
+import com.scalar.db.storage.cassandra.CassandraAdmin.CompactionStrategy;
+import com.scalar.db.storage.cassandra.CassandraAdmin.ReplicationStrategy;
 import com.scalar.db.storage.dynamo.DynamoAdmin;
 import java.io.FileInputStream;
 import java.nio.file.Path;
@@ -27,7 +27,7 @@ public class ConfigFileBasedCommand implements Callable<Integer> {
   private Path configPath;
 
   @Option(
-      names = {"-n", "--network-strategy"},
+      names = {"-n", "--replication-strategy"},
       description =
           "Cassandra network strategy, should be SimpleStrategy or NetworkTopologyStrategy")
   private ReplicationStrategy replicationStrategy;
@@ -35,7 +35,7 @@ public class ConfigFileBasedCommand implements Callable<Integer> {
   @Option(
       names = {"-c", "--compaction-strategy"},
       description = "Cassandra compaction strategy, should be LCS, STCS or TWCS")
-  private CompactStrategy compactStrategy;
+  private CompactionStrategy compactionStrategy;
 
   @Option(
       names = {"-R", "--replication-factor"},
@@ -53,12 +53,14 @@ public class ConfigFileBasedCommand implements Callable<Integer> {
   private Boolean noScaling;
 
   @Option(names = "--no-backup", description = "Disable continuous backup for Dynamo DB")
-  private Boolean dynamoNoBackup;
+  private Boolean noBackup;
+
+  @Option(names = "--coordinator", description = "Create coordinator table", defaultValue = "false")
+  private Boolean coordinator;
 
   @Option(
       names = {"-f", "--schema-file"},
-      description = "Path to schema json file",
-      required = true)
+      description = "Path to schema json file")
   private Path schemaFile;
 
   @Option(
@@ -77,31 +79,38 @@ public class ConfigFileBasedCommand implements Callable<Integer> {
     if (replicationStrategy != null) {
       metaOptions.put(CassandraAdmin.REPLICATION_STRATEGY, replicationStrategy.name());
     }
-    if (compactStrategy != null) {
-      metaOptions.put(CassandraAdmin.COMPACTION_STRATEGY, compactStrategy.name());
+    if (compactionStrategy != null) {
+      metaOptions.put(CassandraAdmin.COMPACTION_STRATEGY, compactionStrategy.name());
     }
     if (replicaFactor != null) {
       metaOptions.put(CassandraAdmin.REPLICATION_FACTOR, replicaFactor);
     }
     if (ru != null) {
-      metaOptions.put("ru", ru);
+      metaOptions.put(DynamoAdmin.REQUEST_UNIT, ru);
     }
     if (noScaling != null) {
-      metaOptions.put("no-scaling", noScaling.toString());
+      metaOptions.put(DynamoAdmin.NO_SCALING, noScaling.toString());
     }
-    if (dynamoNoBackup != null) {
-      metaOptions.put(DynamoAdmin.NO_BACKUP, dynamoNoBackup.toString());
+    if (noBackup != null) {
+      metaOptions.put(DynamoAdmin.NO_BACKUP, noBackup.toString());
     }
 
     DatabaseConfig dbConfig = new DatabaseConfig(new FileInputStream(configPath.toString()));
     SchemaOperator operator = new SchemaOperator(dbConfig);
-    SchemaParser schemaParser = new SchemaParser(schemaFile.toString(), metaOptions);
 
-    if (deleteTables) {
-      operator.deleteTables(schemaParser.getTables());
-    } else {
-      operator.createTables(schemaParser.getTables());
+    if (coordinator) {
+      operator.createCoordinatorTable();
     }
+
+    if (schemaFile != null) {
+      SchemaParser schemaParser = new SchemaParser(schemaFile.toString(), metaOptions);
+      if (deleteTables) {
+        operator.deleteTables(schemaParser.getTables());
+      } else {
+        operator.createTables(schemaParser.getTables());
+      }
+    }
+
     return 0;
   }
 }
